+++ src/pacasam/extractors/bd_ortho_today.py
@@ -13,7 +13,6 @@ dataset_root_path/
 """
 
 
-import os
 from pathlib import Path
 import shutil
 import tempfile